[CCUBE-2114][GZ] move match media mock to _common
